# random-json-server.openapi.yaml
openapi: 3.0.0
info:
  title: Random JSON Server API
  version: 1.0.0
  description: A small HTTP API that returns randomly generated JSON payloads.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /:
    get:
      summary: Get service information
      operationId: getServiceInfo
      responses:
        "200":
          description: Service information and available endpoints.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInfo"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/profile:
    get:
      summary: Get a random user profile
      operationId: getRandomProfile
      responses:
        "200":
          description: A randomly generated user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/weather:
    get:
      summary: Get random weather data
      operationId: getRandomWeather
      responses:
        "200":
          description: Randomly generated weather data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weather"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/metrics:
    get:
      summary: Get random system metrics
      operationId: getRandomMetrics
      responses:
        "200":
          description: Randomly generated system metrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metrics"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orders:
    get:
      summary: Get random order data
      operationId: getRandomOrder
      responses:
        "200":
          description: Randomly generated order data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/inventory:
    get:
      summary: Get random inventory data
      operationId: getRandomInventory
      responses:
        "200":
          description: Randomly generated inventory data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notifications:
    get:
      summary: Get random notification status
      operationId: getRandomNotification
      responses:
        "200":
          description: Randomly generated notification status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    NotFound:
      description: The requested route was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    MethodNotAllowed:
      description: The HTTP method is not supported. Only GET is allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
      additionalProperties: false
    ServiceInfo:
      type: object
      required:
        - message
        - endpoints
      properties:
        message:
          type: string
          example: Simple random JSON server
        endpoints:
          type: array
          items:
            type: string
          example:
            - /api/profile
            - /api/weather
            - /api/metrics
            - /api/orders
            - /api/inventory
            - /api/notifications
      additionalProperties: false
    Profile:
      type: object
      required:
        - id
        - name
        - score
        - active
      properties:
        id:
          type: string
          minLength: 8
          maxLength: 8
          example: ab12cd34
        name:
          type: string
          pattern: "^User-[0-9]{3}$"
          example: User-123
        score:
          type: integer
          minimum: 1
          maximum: 100
          example: 87
        active:
          type: boolean
          example: true
      additionalProperties: false
    Weather:
      type: object
      required:
        - location
        - temperatureF
        - humidity
        - condition
      properties:
        location:
          type: string
          pattern: "^City-[0-9]+$"
          example: City-12
        temperatureF:
          type: integer
          minimum: 40
          maximum: 100
          example: 72
        humidity:
          type: number
          minimum: 0.2
          maximum: 0.9
          example: 0.63
        condition:
          type: string
          enum:
            - sunny
            - cloudy
            - rainy
            - windy
      additionalProperties: false
    Metrics:
      type: object
      required:
        - requestId
        - cpuLoad
        - memoryMb
        - latencyMs
      properties:
        requestId:
          type: string
          minLength: 8
          maxLength: 8
          example: ef56gh78
        cpuLoad:
          type: number
          minimum: 0
          maximum: 1
          example: 0.421
        memoryMb:
          type: integer
          minimum: 128
          maximum: 4096
          example: 1024
        latencyMs:
          type: integer
          minimum: 10
          maximum: 500
          example: 125
      additionalProperties: false
    Order:
      type: object
      required:
        - orderId
        - itemCount
        - totalUsd
        - status
      properties:
        orderId:
          type: string
          pattern: "^ORD-[0-9]{6}$"
          example: ORD-123456
        itemCount:
          type: integer
          minimum: 1
          maximum: 8
          example: 3
        totalUsd:
          type: number
          minimum: 9.99
          maximum: 499.99
          example: 149.95
        status:
          type: string
          enum:
            - processing
            - shipped
            - delivered
            - cancelled
      additionalProperties: false
    Inventory:
      type: object
      required:
        - sku
        - inStock
        - reorderLevel
        - warehouse
      properties:
        sku:
          type: string
          pattern: "^SKU-[0-9]{4}$"
          example: SKU-4821
        inStock:
          type: integer
          minimum: 0
          maximum: 500
          example: 238
        reorderLevel:
          type: integer
          minimum: 10
          maximum: 150
          example: 75
        warehouse:
          type: string
          pattern: "^WH-[1-5]$"
          example: WH-3
      additionalProperties: false
    Notification:
      type: object
      required:
        - notificationId
        - channel
        - priority
        - sent
      properties:
        notificationId:
          type: string
          minLength: 8
          maxLength: 8
          example: z9x8c7v6
        channel:
          type: string
          enum:
            - email
            - sms
            - push
        priority:
          type: string
          enum:
            - low
            - medium
            - high
        sent:
          type: boolean
          example: true
      additionalProperties: false